+++ permissions/src/epilot/models/shared/grantcondition.py
@@ -7,17 +7,17 @@ from enum import Enum
 from epilot import utils
 from typing import Any
 
-class EqualsConditionOperation(str, Enum):
+class GrantConditionOperation(str, Enum):
     EQUALS = 'equals'
 
 
 @dataclass_json(undefined=Undefined.EXCLUDE)
 
 @dataclasses.dataclass
-class EqualsCondition:
+class GrantCondition:
     r"""Check if attribute equals to any of the values"""
     attribute: str = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('attribute') }})
-    operation: EqualsConditionOperation = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('operation') }})
+    operation: GrantConditionOperation = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('operation') }})
     values: list[Any] = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('values') }})